+++ backend/src/StockIntel.Application/Services/UserService.cs
@@ -1,7 +1,9 @@
  using StockIntel.Application.Interfaces.Services;
  using StockIntel.Domain.Entities;
  using StockIntel.Application.Common.Interfaces;
-using StockIntel.Application.DTOs.User;
+ using StockIntel.Application.DTOs.User;
+ using StockIntel.Application.Exceptions;
+
  namespace StockIntel.Application.Services;
 
  public class UserService : IUserService
@@ -20,8 +22,10 @@ using StockIntel.Application.DTOs.User;
 {
     var user = await _userRepository.GetByIdAsync(id);
 
-    if (user == null)
-        return null;
+   if (user == null)
+    {
+        throw new NotFoundException("User not found.");
+    }
 
     return new UserDto
     {
@@ -41,7 +45,9 @@ using StockIntel.Application.DTOs.User;
     var user = await _userRepository.GetByEmailAsync(email);
 
     if (user == null)
-        return null;
+    {
+        throw new NotFoundException("User not found.");
+    }
 
     return new UserDto
     {
@@ -72,17 +78,40 @@ using StockIntel.Application.DTOs.User;
 
     public async Task AddAsync (User user)
     {
+        var existingUser = await _userRepository.GetByEmailAsync(user.Email);
+
+        if (existingUser != null)
+        {
+            throw new BadRequestException("A user with this email already exists.");
+        }
+        if (string.IsNullOrWhiteSpace(user.FirstName))
+            throw new BadRequestException("First name is required.");
+
+        if (string.IsNullOrWhiteSpace(user.LastName))
+            throw new BadRequestException("Last name is required.");
         await _userRepository.AddAsync(user);
         await _unitOfWork.SaveChangesAsync();
     }
 
     public async Task UpdateAsync(Guid id, User user) 
     {
+        var existingUser = await _userRepository.GetByIdAsync(id);
+
+        if (existingUser == null)
+        {
+            throw new NotFoundException("User not found.");
+        }
         await _userRepository.UpdateAsync(id, user);
         await _unitOfWork.SaveChangesAsync();
     }
     public async Task DeleteAsync(Guid id)
     {
+        var existingUser = await _userRepository.GetByIdAsync(id);
+
+        if (existingUser == null)
+        {
+            throw new NotFoundException("User not found.");
+        }
         await _userRepository.DeleteAsync(id);
         await _unitOfWork.SaveChangesAsync();
     }
